Moving edge index consistency test to generator_tests_support.
