+++ routing/routing_integration_tests/edge_index_consistency_tests.cpp
@@ -1,0 +1,14 @@
+#include "testing/testing.hpp"
+
+#include "generator/generator_tests_support/test_mwm_consistency.hpp"
+
+namespace
+{
+using namespace generator;
+using namespace generator::tests_support;
+
+UNIT_TEST(EdgeIndexConsistency_MoscowTest)
+{
+  Index index;
+}
+}  // namespace
